Updated the theme and wedding menus!
Today is my second wedding anniversary, and I think it's auspicious that
I make changes to these menus on this day. Much like our wedding was
notable for its food, so too are these menus. Granted, ours was mostly
the world's fastest-assembled taco bar and some Rosatti's, but man you
don't know. You weren't there. I've seen some shit.

Also the themed menus changed. Seafood is out this year, which seems
like a questionable decision considering there's about to be a heck of a
lot more of it when the ice caps melt. The price we pay to get our
revenge for the Titanic, I guess.