Thanksgiving is approaching!
I know Sakurai said everyone was going to be in Smash, but I didn't
think he meant major American holidays too. Rumors are flying about
what the moveset is going to look like, but I think we can all agree
that the matchups with Kirby are going to be pretty unbalanced.

Dude's just so good at eating!