Christmas is over!
As we enter the winter of our discontent, let us all take a moment to meditate on what it is
we're so discontented with. Is it the amount of family-friendly network programming? Is it
the rising cost of basic staples like eggs and milk? Is it that despite our best efforts, we still
haven't cracked the secrets of flying jet shoes? Or are we all actually mostly fine? Probably
that last one.